+++ tests/eslint/eslint.config.mjs
@@ -2,6 +2,7 @@ import js from "@eslint/js"
 import tseslint from "typescript-eslint"
 import reactPlugin from "eslint-plugin-react"
 import reactHooksPlugin from "eslint-plugin-react-hooks"
+import globals from "globals"
 
 export default tseslint.config(
   js.configs.recommended,
@@ -13,11 +14,17 @@ export default tseslint.config(
       "react-hooks": reactHooksPlugin,
     },
     languageOptions: {
+      globals: {
+        ...globals.node,
+        ...globals.browser,
+      },
       parser: tseslint.parser,
       parserOptions: {
         ecmaFeatures: {
           jsx: true,
         },
+        tsconfigRootDir: process.cwd(),
+        project: ["./tsconfig.json"],
       },
     },
     rules: {